+++ tests/projects/c++/modules/dependence/CMakeLists.txt
@@ -1,0 +1,43 @@
+# this is the build file for project 
+# it is autogenerated by the xmake build system.
+# do not edit by hand.
+
+# project
+cmake_minimum_required(VERSION 3.15.0)
+cmake_policy(SET CMP0091 NEW)
+project(dependence LANGUAGES CXX C)
+
+# target
+add_executable(dependence "")
+set_target_properties(dependence PROPERTIES OUTPUT_NAME "dependence")
+set_target_properties(dependence PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_SOURCE_DIR}/build/macosx/x86_64/release")
+target_compile_options(dependence PRIVATE
+    $<$<COMPILE_LANGUAGE:C>:-m64>
+    $<$<COMPILE_LANGUAGE:CXX>:-m64>
+    $<$<COMPILE_LANGUAGE:C>:-isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX11.3.sdk>
+    $<$<COMPILE_LANGUAGE:CXX>:-isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX11.3.sdk>
+    $<$<COMPILE_LANGUAGE:CXX>:-fmodules>
+    $<$<COMPILE_LANGUAGE:CXX>:-fbuiltin-module-map>
+    $<$<COMPILE_LANGUAGE:CXX>:-fimplicit-modules>
+    $<$<COMPILE_LANGUAGE:CXX>:-fno-implicit-module-maps>
+)
+target_compile_features(dependence PRIVATE cxx_std_20)
+if(MSVC)
+    set_property(TARGET dependence PROPERTY
+        MSVC_RUNTIME_LIBRARY "MultiThreaded$<$<CONFIG:Debug>:Debug>")
+endif()
+target_link_options(dependence PRIVATE
+    -m64
+    -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX11.3.sdk
+)
+target_sources(dependence PRIVATE
+    src/hello_impl.cpp
+    src/mod_impl.cpp
+    src/main.cpp
+    src/hello_impl.cpp
+    src/mod_impl.cpp
+    src/main.cpp
+    src/mod.mpp
+    src/hello.mpp
+)
+
